requirements.txt
streamlit>=1.32.0
anthropic>=0.40.0
python-dotenv>=1.0.0
Pillow>=10.0.0
reportlab>=4.0.0
google-generativeai>=0.8.0
